+++ modules/config.js
@@ -32,3 +32,5 @@ AXIOM.damageTypes = {
   necrotic: "Necrotic",
   psychic: "Psychic",
 };
+
+